# internal/httpserver/swaggerdocs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
...
        "title": "GoDNS API",
        "contact": {
            "name": "GoDNS Support",
            "url": "https://github.com/rogerwesterbo/godns",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/rogerwesterbo/godns/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:14000",
    "basePath": "/",
    "paths": {
        "/api/v1/admin/cache/clear": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Clear all entries from the DNS response cache",
                "tags": [
                    "Admin"
                ],
                "summary": "Clear DNS cache",
                "responses": {
                    "204": {
                        "description": "Cache cleared successfully"
                    },
                    "503": {
                        "description": "Cache service not available",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/admin/cache/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Get detailed statistics about DNS response caching",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get cache statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_httpserver_handlers_v1adminhandler.CacheStats"
                        }
                    }
                }
            }
        },
...
        "/api/v1/admin/loadbalancer/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Get detailed statistics about load balancing including backend health status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get load balancer statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_httpserver_handlers_v1adminhandler.LoadBalancerStats"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/querylog/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Get statistics about DNS queries including cache hit rate and blocked queries",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get query log statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_httpserver_handlers_v1adminhandler.QueryLogStats"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/ratelimiter/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Get statistics about rate limiting including active limiters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get rate limiter statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_httpserver_handlers_v1adminhandler.RateLimiterStats"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Get comprehensive statistics about DNS caching, rate limiting, load balancing, health checks, and query logging",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get system statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_httpserver_handlers_v1adminhandler.SystemStats"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_httpserver_handlers_v1adminhandler.CacheStats": {
            "type": "object",
            "properties": {
                "current_size": {
                    "type": "integer"
                },
                "enabled": {
                    "type": "boolean"
                },
                "hit_rate": {
                    "type": "string"
                },
                "max_size": {
                    "type": "integer"
                },
                "ttl_minutes": {
                    "type": "integer"
                }
            }
        },
        "internal_httpserver_handlers_v1adminhandler.HealthCheckResultInfo": {
            "type": "object",
            "properties": {
                "error_message": {
                    "type": "string"
                },
                "healthy": {
                    "type": "boolean"
                },
                "last_check": {
                    "type": "string"
                },
                "last_failure": {
                    "type": "string"
                },
                "last_success": {
                    "type": "string"
                },
                "target": {
                    "type": "string"
                }
            }
        },
        "internal_httpserver_handlers_v1adminhandler.HealthCheckStats": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "interval_seconds": {
                    "type": "integer"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_httpserver_handlers_v1adminhandler.HealthCheckResultInfo"
                    }
                },
                "timeout_seconds": {
                    "type": "integer"
                },
                "total_targets": {
                    "type": "integer"
                }
            }
        },
        "internal_httpserver_handlers_v1adminhandler.LoadBalancerBackendInfo": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "healthy": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                },
                "weight": {
                    "type": "integer"
                }
            }
        },
        "internal_httpserver_handlers_v1adminhandler.LoadBalancerStats": {
            "type": "object",
            "properties": {
                "backend_groups": {
                    "type": "integer"
                },
                "backends": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_httpserver_handlers_v1adminhandler.LoadBalancerBackendInfo"
                    }
                },
                "enabled": {
                    "type": "boolean"
                },
                "healthy_backends": {
                    "type": "integer"
                },
                "strategy": {
                    "type": "string"
                },
                "total_backends": {
                    "type": "integer"
                }
            }
        },
        "internal_httpserver_handlers_v1adminhandler.QueryLogStats": {
            "type": "object",
            "properties": {
                "blocked_queries": {
                    "type": "integer"
                },
                "blocked_rate": {
                    "type": "number"
                },
                "cache_hit_rate": {
                    "type": "number"
                },
                "cached_queries": {
                    "type": "integer"
                },
                "enabled": {
                    "type": "boolean"
                },
                "total_queries": {
                    "type": "integer"
                }
            }
        },
        "internal_httpserver_handlers_v1adminhandler.RateLimiterStats": {
            "type": "object",
            "properties": {
                "active_limiters": {
                    "type": "integer"
                },
                "burst": {
                    "type": "integer"
                },
                "enabled": {
                    "type": "boolean"
                },
                "qps": {
                    "type": "integer"
                }
            }
        },
        "internal_httpserver_handlers_v1adminhandler.SystemStats": {
            "type": "object",
            "properties": {
                "cache": {
                    "$ref": "#/definitions/internal_httpserver_handlers_v1adminhandler.CacheStats"
                },
                "health_check": {
                    "$ref": "#/definitions/internal_httpserver_handlers_v1adminhandler.HealthCheckStats"
                },
                "load_balancer": {
                    "$ref": "#/definitions/internal_httpserver_handlers_v1adminhandler.LoadBalancerStats"
                },
                "query_log": {
                    "$ref": "#/definitions/internal_httpserver_handlers_v1adminhandler.QueryLogStats"
                },
                "rate_limiter": {
                    "$ref": "#/definitions/internal_httpserver_handlers_v1adminhandler.RateLimiterStats"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "http://localhost:14101/realms/godns/protocol/openid-connect/token"
        }
    },
...
}